fix(horeca): permitir vender por kg/L (decimales) y borrar recetas de prueba
- ItemComanda.cantidad pasa de Integer a BigDecimal (igual que el resto del
  sistema) para poder vender fracciones reales de inventario (2.5 kg de
  harina, 1.5 L de aceite), no solo unidades enteras. Bug real: el buscador
  de Venta Rapida truncaba a entero con parseInt y limitaba min=1, por lo que
  no dejaba vender ni ver bien el stock de articulos por peso/volumen.
- Nuevo DELETE /api/horeca/escandallos/{id} + boton de borrar en cada receta
  de la pantalla Recetas & Escandallo, para poder limpiar datos de prueba
  (rechaza el borrado con un mensaje claro si la receta ya tiene ventas).

+++ src/main/java/com/auroraplus/core/sync/SincronizacionLoteService.java
@@ -117,7 +117,7 @@ public class SincronizacionLoteService {
         public Long escandalloId;
         public String nombrePlato;
         public String estacionCocina;
-        public Integer cantidad;
+        public BigDecimal cantidad;
         public BigDecimal precioUnitario;
     }
